+++ examples/flutter/bin/check_config.dart
@@ -27,7 +27,7 @@ const _configJson = '''
 
 void main(List<String> args) {
   final libraryPath = args.isEmpty ? null : args.first;
-  final singbox = SingboxFfi.openDefault(libraryPath);
+  final singbox = SingboxFfi.openBundled(libraryPath);
 
   print('sing-box version: ${singbox.version()}');
   print('go version: ${singbox.goVersion()}');